requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
kaleido==0.2.1
multidict==6.0.4
nextcord==2.4.2
numpy==1.24.3
packaging==23.1
pandas==2.0.1
plotly==5.14.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tenacity==8.2.2
typing-extensions==4.5.0
tzdata==2023.3
yarl==1.9.2
